Wed, Aug 16, 2023, 1:03 PM +01:00

+++ app.py
@@ -71,7 +71,7 @@ def new():
 @login_required
 def delete():
     """delete project"""
-    db.execute("DELETE FROM progressions WHERE id = ? AND user_id = ?", request.form.get("id"), )
+    db.execute("DELETE FROM progressions WHERE id = ? AND user_id = ?", request.form.get("id"), session.get("user_id"))
     return redirect("/")
